Don't rely on position or number of reserved fields.
The way we handled reserved fields so far was BAD for forward compatibility,
because we referenced the name "reserved", which may disappear in future
versions.
Or even worse, the offset of `reserved` may move across versions, in which
case we may be writing past the struct bounds!

Instead, we now memset() the entire struct to 0, and only write non-reserved
fields, as the lz4 docs describe.